undo insertion of defaults for empty template components
