+++ src/components/SegmentTemplate.tsx
