--- src/template/types/BarAndDecContainerType.ts
@@ -1,13 +1,0 @@
-import type { BarElementType } from "./BarElementType";
-
-type BarAndDecContainerType =
-{
-  type: "bar-dec-container";
-  elements: BarElementType[];
-  CSS?: string;
-  decorationWidth?: string;
-  order?: number;
-  onClickHandler?: React.MouseEventHandler<HTMLDivElement>;
-}
-
-export type { BarAndDecContainerType };
