--- src/template/types/BarContentContainerElementType.ts
@@ -1,8 +1,0 @@
-import type { BarAndDecContainerType } from "./BarAndDecContainerType";
-import type { DecorationType } from "./DecorationType";
-
-type BarContentContainerElementType =
-| BarAndDecContainerType
-| DecorationType;
-
-export type { BarContentContainerElementType };
