--- src/template/types/BarElementType.ts
@@ -1,8 +1,0 @@
-import type { BarType } from './BarType';
-import type { DecorationType } from './DecorationType';
-
-type BarElementType =
-| BarType
-| DecorationType;
-
-export type { BarElementType };
