--- src/template/types/FullBarElementType.ts
@@ -1,8 +1,0 @@
-import type { BarContentContainerType } from "./BarContentContainerType"
-import type { DecorationType } from "./DecorationType"
-
-type FullBarElementType =
-| BarContentContainerType
-| DecorationType
-
-export type { FullBarElementType };
