--- src/template/types/PlotCSS.ts
@@ -1,6 +1,0 @@
-type PlotCSS = {
-    "bar-plot": string;
-    [key: string]: string;
-}
-
-export type { PlotCSS };
