--- src/template/types/Vars.ts
@@ -1,5 +1,0 @@
-type Vars = {
-    [name: string]: (string | string[])[] | (number | number[])[],
-}
-
-export type { Vars };
